make EVERYTHING outlined!
Previously, we used the elevated material styles for all mat-card
elements. This isn't even mentioned in angular's documentation anymore,
as there they default to, and only show the `outline` or `outlined`
style (angular-material is inconsistent with this)

Now, every mat-card I could find has `appearance="outlined"` set on it
to achieve the desired look. The sidebar/`mat-drawer` was also adjusted
to fit.

I swear there was a way to set a default appearance in angular-material,
but I can't for the life of me find it again.